assessment/quiz/requirements.txt
streamlit
pymongo
python-dotenv
fireworks-ai